refactor: reorganize game routing and rename components for game lobby structure
