+++ src/App.tsx
@@ -27,10 +27,11 @@ import {
   ProfilePage,
   FriendPage,
   ChatRoomPage,
-  GameSelectPage,
   TournamentPage,
-  GameAutoMatchingPage,
   OAuthCallbackPage,
+  GamePage,
+  GameLobbyPage,
+  GameAutoMatchingPage,
   CustomMatchingPage,
 } from '@/pages';
 
@@ -94,10 +95,13 @@ const App = () => {
         { path: PATH.PROFILE, element: <ProfilePage /> },
         { path: PATH.FRIEND, element: <FriendPage /> },
         { path: PATH.FRIEND_CHATROOM, element: <ChatRoomPage /> },
-        { path: PATH.GAME_SELECT, element: <GameSelectPage /> },
-        { path: PATH.TOURNAMENT, element: <TournamentPage /> },
-        { path: PATH.GAME_AUTO_MATCHING, element: <GameAutoMatchingPage /> },
-        { path: PATH.GAME_CUSTOM_MATCHING, element: <CustomMatchingPage /> },
+
+        { path: PATH.GAME, element: <GamePage /> },
+
+        { path: PATH.GAME_LOBBY, element: <GameLobbyPage /> },
+        { path: PATH.GAME_LOBBY_TOURNAMENT, element: <TournamentPage /> },
+        { path: PATH.GAME_LOBBY_AUTO_MATCHING, element: <GameAutoMatchingPage /> },
+        { path: PATH.GAME_LOBBY_CUSTOM_MATCHING, element: <CustomMatchingPage /> },
       ],
     },
   ];
