+++ src/pages/GameLobbyPage/AutoMatchingPage/Auto1vs1/styles.css.ts
