+++ src/pages/GameLobbyPage/AutoMatchingPage/AutoTournament/index.tsx
