+++ src/pages/GameLobbyPage/AutoMatchingPage/AutoTournament/styles.css.ts
