+++ src/pages/GameLobbyPage/AutoMatchingPage/index.tsx
