+++ src/pages/GameLobbyPage/CustomMatchingPage/Custom1vs1/styles.css.ts
