+++ src/pages/GameLobbyPage/CustomMatchingPage/CustomTournament/styles.css.ts
