+++ src/pages/GameLobbyPage/_components/invite-friend-dialog/index.tsx
