+++ src/pages/GameLobbyPage/_components/invite-friend-dialog/styles.css.ts
