+++ src/pages/GameLobbyPage/_components/user-card/styles.css.ts
