+++ src/pages/GameLobbyPage/_components/waiting-message/index.tsx
