+++ src/pages/GameLobbyPage/_components/waiting-message/styles.css.ts
