+++ src/pages/GamePage/index.tsx
@@ -1,0 +1,3 @@
+export const GamePage = () => {
+  return <></>;
+};
